1. improve object management 2. statements are now harvested either when connections are cosed or 3. when row fetch has completed and a new operation is fired at driver to free up statement. 4. prepared statements will live as long as required and can be re-submitted with new parameters. - not yet done.

+++ src/OdbcStatement.cpp
@@ -36,6 +36,7 @@ namespace mssql
 
 	OdbcStatement::~OdbcStatement()
 	{
+		//fprintf(stderr, "destruct OdbcStatement\n");
 		if (statement) {
 			statement.Free();
 		}
@@ -44,7 +45,7 @@ namespace mssql
 		}
 	}
 
-	OdbcStatement::OdbcStatement(int statementId, OdbcConnectionHandle &c)
+	OdbcStatement::OdbcStatement(size_t statementId, OdbcConnectionHandle &c)
 		:	
 		connection(c),
 		error(nullptr),
